reduce non-timed-out history lifetime to 1 day
The maximum number of time that the official Signal clients will listen
to a remote delete request is one day from the server receive timestamp,
so keeping such a history for more than 1 day isn't necessary. However,
we continue to retain history of LaTeX requests that took too long for
10 days for abuse monitoring purposes. Also, the server logs (journald)
still retains logs for 10 days.